+++ domain/src/test/java/com/github/dodobest/domain/CalculatorTest.kt
@@ -1,6 +1,5 @@
-package com.github.dodobest
+package com.github.dodobest.domain
 
-import com.github.dodobest.domain.Calculator
 import com.google.common.truth.Truth.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
